pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "betteredit"
version = "0.1.0"
description = "Analyze visual weight, flow, and aesthetics of images"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.8"
authors = [
  { name = "Aashi Gupta", email = "aashi.gupta2@icloud.com" },
]
dependencies = [
  # Core dependencies - RESOLVED: Use only opencv-contrib-python
  "opencv-contrib-python>=4.8.0,<5.0.0",
  
  # Scientific computing (actually used)
  "numpy>=1.23",
  "scipy>=1.7.0",
  "scikit-learn>=1.2",
  "Pillow>=9.0.0",
  
  # Visualization (actually used)
  "matplotlib>=3.7",
  
  # Configuration and utilities (actually used)
  "PyYAML>=5.4",
  "pydantic>=2.0",
  "pydantic-settings>=2.0",
  "loguru",
  
  # Neural inter-fusion dependencies
  "torch>=2.0.0",
  "torchvision>=0.15.0",
  
  # Development and testing
  "pytest",
  "mypy",
  "types-PyYAML",
]

[project.urls]
Homepage = "https://github.com/isaashi/betteredit"
Repository = "https://github.com/isaashi/betteredit"

[tool.setuptools.packages.find]
where = ["src"]

[project.entry-points."console_scripts"]
betteredit = "betteredit.cli:main"